guardrail/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
openai>=1.0.0
faiss-cpu>=1.7.0
numpy
pydantic>=2.0.0
